+++ apps/www/src/app/(main)/(docs)/components/chip/examples/variants.tsx
@@ -2,8 +2,7 @@
 
 "use client";
 
-import { Chip, UIStylesProvider, type ChipProps } from "@jamsr-ui/react";
-import { cn } from "@jamsr-ui/utils";
+import { Chip, type ChipProps } from "@jamsr-ui/react";
 
 export const ChipVariants = () => {
   const variants: ChipProps["variant"][] = ["solid", "outlined", "flat", "dot"];
@@ -14,28 +13,6 @@ export const ChipVariants = () => {
           {variant}
         </Chip>
       ))}
-
-      <UIStylesProvider
-        chip={{
-          props({ variant, color }) {
-            return variant === "dot"
-              ? {
-                  classNames: {
-                    base: cn(
-                      "border-2 border-divider",
-                      color === "danger" && "border-danger",
-                    ),
-                  },
-                }
-              : {};
-          },
-        }}
-      >
-        <Chip variant="dot">Hello</Chip>
-        <Chip variant="dot" color="danger">
-          Hello
-        </Chip>
-      </UIStylesProvider>
     </div>
   );
 };