+++ apps/www/src/app/(main)/(docs)/components/input/examples/variants.tsx
@@ -4,7 +4,8 @@ export const InputVariants = () => {
   return (
     <div className="space-y-4">
       <Input label="Outlined" variant="outlined" />
-      <Input label="Standard" variant="standard" />
+      <Input label="Standard" variant="standard" placeholder="Standard Input" />
+      <Input label="Underlined" variant="underlined" placeholder="Underlined Input" />
     </div>
   );
 };