+++ packages/components/avatar/src/styles.ts
@@ -2,7 +2,7 @@ import { type VariantProps, radiusBaseVariant, tv } from "@jamsr-ui/utils";
 
 export const test = "";
 export const avatarVariants = tv({
-  base: "flex flex-shrink-0 items-center justify-center overflow-hidden rounded-full bg-content1 object-cover text-md font-medium [&>img]:object-cover",
+  base: "flex flex-shrink-0 items-center justify-center overflow-hidden rounded-full bg-content1 object-cover text-md font-medium [&>img]:size-full [&>img]:object-cover",
   variants: {
     size: {
       xs: "size-6",