+++ packages/components/select/src/select-item.tsx
@@ -3,7 +3,12 @@ import { useHover, useMergeRefs } from "@jamsr-ui/hooks";
 import { CheckIcon } from "@jamsr-ui/shared-icons";
 import { useUIStyle } from "@jamsr-ui/styles";
 import type { ComponentPropsWithAs, UIProps } from "@jamsr-ui/utils";
-import { dataAttr, deepMergeProps, mergeGlobalProps } from "@jamsr-ui/utils";
+import {
+  cn,
+  dataAttr,
+  deepMergeProps,
+  mergeGlobalProps,
+} from "@jamsr-ui/utils";
 import { useSelectContext } from "./use-select-context";
 
 type Props = {
@@ -38,16 +43,13 @@ export const SelectItem = <T extends React.ElementType = "button">(
     activeIndex,
     getItemProps,
     handleSelect,
-    setValue,
-    isMultiple,
     value: values,
     styles,
+    classNames,
+    onSelectValue,
   } = useSelectContext();
 
-  // eslint-disable-next-line no-nested-ternary
-  const listLabel =
-    label ?? (typeof children === "string" ? children : undefined);
-
+  const listLabel = label ?? (typeof children === "string" ? children : value);
   if (!listLabel) {
     console.warn(`No label provided for list item with value ${value}`);
   }
@@ -63,25 +65,13 @@ export const SelectItem = <T extends React.ElementType = "button">(
   const refs = useMergeRefs([ref, disableRef]);
 
   const handleClick = () => {
-    const getNewValue = (): string[] => {
-      if (isMultiple) {
-        const $value = new Set(values);
-        if ($value.has(value)) {
-          $value.delete(value);
-          return [...$value];
-        }
-        $value.add(value);
-        return [...$value];
-      }
-      return [value];
-    };
-    setValue(getNewValue());
+    onSelectValue(value);
     handleSelect(index);
   };
 
   const Component = as ?? "li";
   const style = styles.selectItem({
-    className,
+    className: cn(classNames?.selectItem, className),
   });
 
   const handleOnKeyDown = (e: React.KeyboardEvent<HTMLButtonElement>) => {